+++ projects/mk-kit/directives/directives/index.ts
@@ -7,3 +7,4 @@ export * from './intersect';
 export * from './infinite-scroll';
 export * from './ripple';
 export * from './mask';
+export * from './field';